Perform some refactoring and code commenting.
- renamed PromptsDialog to FileUploadDialog.
